Add files via upload
Histogram_MDV.py : Histogram of magnitude of the difference in velocities (cell-neighborhood) (log-scale)
Histogram_MDV_xy.py : Histogram of magnitude of the difference in velocities in x, y (cell-neighborhood) (log-scale)
Histogram_MDVN.py : Histogram of magnitude of the difference in velocities (cell-neighborhood), normalized to the magnitude of the velocity of the neighborhood (log-scale)
plot_magdvelocity_time.py: Plot of magnitude of the difference in velocities (cell-neighborhood) vs time. (One point for cell)
plot_magdvelocity_distcentro.py: Plot of magnitude of the difference in velocities (cell-neighborhood) vs the distance of the cell to the center. (One point for cell)
plot_magdvelocity_2D_xy.py: Diagram of magnitude of the difference in velocities (cell-neighborhood) with different colors for different magnitude. For each time.